spatial scene parts are called parts again
