+++ crates/re_space_view_spatial/src/scene/parts/arrows3d.rs
