+++ crates/re_space_view_spatial/src/scene/parts/boxes3d.rs
