+++ crates/re_space_view_spatial/src/scene/parts/lines3d.rs
