+++ crates/re_space_view_spatial/src/scene/parts/points3d.rs
@@ -1,5 +1,5 @@
 use re_components::{
-    ClassId, ColorRGBA, Component, InstanceKey, KeypointId, Label, Point2D, Radius,
+    ClassId, ColorRGBA, Component as _, InstanceKey, KeypointId, Label, Point3D, Radius,
 };
 use re_data_store::{EntityPath, InstancePathHash};
 use re_log_types::ComponentName;
@@ -8,23 +8,21 @@ use re_viewer_context::{ResolvedAnnotationInfo, SceneQuery, SpaceViewHighlights,
 
 use crate::scene::{
     contexts::{SpatialSceneContext, SpatialSceneEntityContext},
-    load_keypoint_connections,
-    spatial_scene_part::{SpatialScenePart, SpatialScenePartData},
-    UiLabel, UiLabelTarget,
+    load_keypoint_connections, UiLabel, UiLabelTarget,
 };
 
 use super::{
     instance_key_to_picking_id, instance_path_hash_for_picking, process_annotations_and_keypoints,
-    process_colors, process_radii,
+    process_colors, process_radii, SpatialScenePart, SpatialScenePartData,
 };
 
-pub struct Points2DScenePart {
+pub struct Points3DPart {
     /// If the number of points in the batch is > max_labels, don't render point labels.
     pub max_labels: usize,
     pub data: SpatialScenePartData,
 }
 
-impl Default for Points2DScenePart {
+impl Default for Points3DPart {
     fn default() -> Self {
         Self {
             max_labels: 10,
@@ -33,12 +31,13 @@ impl Default for Points2DScenePart {
     }
 }
 
-impl Points2DScenePart {
+impl Points3DPart {
     fn process_labels<'a>(
-        entity_view: &'a EntityView<Point2D>,
+        entity_view: &'a EntityView<Point3D>,
         instance_path_hashes: &'a [InstancePathHash],
         colors: &'a [egui::Color32],
         annotation_infos: &'a [ResolvedAnnotationInfo],
+        world_from_obj: glam::Affine3A,
     ) -> Result<impl Iterator<Item = UiLabel> + 'a, QueryError> {
         let labels = itertools::izip!(
             annotation_infos.iter(),
@@ -54,7 +53,9 @@ impl Points2DScenePart {
                     (Some(point), Some(label)) => Some(UiLabel {
                         text: label,
                         color: *color,
-                        target: UiLabelTarget::Point2D(egui::pos2(point.x, point.y)),
+                        target: UiLabelTarget::Position3D(
+                            world_from_obj.transform_point3(point.into()),
+                        ),
                         labeled_instance: *labeled_instance,
                     }),
                     _ => None,
@@ -67,7 +68,7 @@ impl Points2DScenePart {
     fn process_entity_view(
         &mut self,
         query: &SceneQuery<'_>,
-        ent_view: &EntityView<Point2D>,
+        ent_view: &EntityView<Point3D>,
         ent_path: &EntityPath,
         ent_context: &SpatialSceneEntityContext<'_>,
     ) -> Result<(), QueryError> {
@@ -103,18 +104,14 @@ impl Points2DScenePart {
                 &instance_path_hashes_for_picking,
                 &colors,
                 &annotation_infos,
+                ent_context.world_from_obj,
             )?);
         }
 
         {
             let mut point_builder = ent_context.shared_render_builders.points();
             let point_batch = point_builder
-                .batch("2d points")
-                .depth_offset(ent_context.depth_offset)
-                .flags(
-                    re_renderer::renderer::PointCloudBatchFlags::FLAG_DRAW_AS_CIRCLES
-                        | re_renderer::renderer::PointCloudBatchFlags::FLAG_ENABLE_SHADING,
-                )
+                .batch("3d points")
                 .world_from_obj(ent_context.world_from_obj)
                 .outline_mask_ids(ent_context.highlight.overall)
                 .picking_object_id(re_renderer::PickingLayerObjectId(ent_path.hash64()));
@@ -123,7 +120,7 @@ impl Points2DScenePart {
                 re_tracing::profile_scope!("collect_points");
                 ent_view
                     .iter_primary()?
-                    .filter_map(|pt| pt.map(glam::Vec2::from))
+                    .filter_map(|pt| pt.map(glam::Vec3::from))
             };
 
             let picking_instance_ids = ent_view.iter_instance_keys().map(|instance_key| {
@@ -133,8 +130,7 @@ impl Points2DScenePart {
                     ent_context.highlight.any_selection_highlight,
                 )
             });
-
-            let mut point_range_builder = point_batch.add_points_2d(
+            let mut point_range_builder = point_batch.add_points(
                 ent_view.num_instances(),
                 point_positions,
                 radii,
@@ -159,17 +155,17 @@ impl Points2DScenePart {
                     }
                 }
             }
-        };
+        }
 
         load_keypoint_connections(ent_context, ent_path, keypoints);
 
         {
-            re_tracing::profile_scope!("points2d.bounding_box");
+            re_tracing::profile_scope!("points3d.bounding_box");
             self.data.bounding_box = self.data.bounding_box.union(
                 macaw::BoundingBox::from_points(
                     ent_view
                         .iter_primary()?
-                        .filter_map(|pt| pt.map(|pt| glam::vec3(pt.x, pt.y, 0.0))),
+                        .filter_map(|pt| pt.map(|pt| pt.into())),
                 )
                 .transform_affine3(&ent_context.world_from_obj),
             );
@@ -179,12 +175,12 @@ impl Points2DScenePart {
     }
 }
 
-impl SpatialScenePart<7> for Points2DScenePart {
-    type Primary = Point2D;
+impl SpatialScenePart<7> for Points3DPart {
+    type Primary = Point3D;
 
     fn archetype() -> [ComponentName; 7] {
         [
-            Point2D::name(),
+            Point3D::name(),
             InstanceKey::name(),
             ColorRGBA::name(),
             Radius::name(),
@@ -201,7 +197,7 @@ impl SpatialScenePart<7> for Points2DScenePart {
         context: &SpatialSceneContext,
         highlights: &SpaceViewHighlights,
     ) -> Vec<re_renderer::QueueableDrawData> {
-        re_tracing::profile_scope!("Points2DPart");
+        re_tracing::profile_scope!("Points3DPart");
 
         Self::for_each_entity_view(
             ctx,
