+++ crates/re_space_view_spatial/src/scene/parts/spatial_scene_part.rs
@@ -5,7 +5,7 @@ use re_query::{query_primary_with_history, EntityView, QueryError};
 use re_renderer::DepthOffset;
 use re_viewer_context::{ScenePartImpl, SceneQuery, SpaceViewHighlights, ViewerContext};
 
-use super::{
+use crate::scene::{
     contexts::{SpatialSceneContext, SpatialSceneEntityContext},
     UiLabel,
 };
